+++ src/app/(afterLogin)/(main)/daily/[day]/page.tsx
@@ -1,3 +1,12 @@
+import UpperTabBar from '@/app/_components/upperTabBar/UpperTabBar';
+
 export default function DailyPage() {
-  return <div />;
+  return (
+    <>
+      <div className="mt-8 mx-16">
+        <UpperTabBar />
+      </div>
+      <div />
+    </>
+  );
 }
